api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "allure-pytest>=2.14.0",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "influxdb>=5.3.2",
    "paho-mqtt>=2.1.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-tornasync>=0.6.0.post2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "sse-starlette>=2.2.1",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]
